% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data()
}
\value{
dataframe
}
\description{
Get data from http://www.mca.gov.cn/article/sj/xzqh/2019/2019/201911250933.html
}
\examples{
\donttest{
  get_data()
}

}
